-removed breeze dependency -renamed and moved the folder to saul-examples -removed plotting -added a short readme
