+++ build.sbt
@@ -69,10 +69,7 @@ lazy val commonSettings = Seq(
     "com.gurobi" % "gurobi" % "6.0",
     "org.apache.commons" % "commons-math3" % "3.0",
     "org.scalatest" % "scalatest_2.11" % "2.2.4",
-    "ch.qos.logback" % "logback-classic" % "1.1.7",
-    "org.scalanlp" %% "breeze" % "0.12",
-    "org.scalanlp" %% "breeze-natives" % "0.12",
-    "org.scalanlp" %% "breeze-viz" % "0.12"
+    "ch.qos.logback" % "logback-classic" % "1.1.7"
   ),
   fork := true,
   connectInput in run := true,
